mctpd: endpoint_send_set_endpoint_id should only send a Set Endpoint ID command
An earlier change modified endpoint_send_set_endpoint_id() to also
modify the peer state. That's not really the intention of this function
- we should only be performing the command - hence the const peer pointer.

Instead, have the requested pool size returned as an out arg, for the
caller to handle. This means we can keep all of the pool logic in the
one place in the caller.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>