mctpd: fix EID allocation when only single-EID ranges are present
While walking through the available EID space, we update the 'best-fit'
allocation if the current extent is larger than the current best-fit.

However, if the only remaining allocations all have an extent of zero,
we will never populate the best-fit range.

Add a start case for the best-fit: the first single-eid allocation we
find.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>